Mappings: Move Rhino mappings provider to separate file
This is a bit cleaner, and keep the Mapping class nice and focused
